Create a reusable `optionsDialog` component
